UI tweaks

+++ game_logic.h
@@ -62,4 +62,4 @@ void moveHorizontal(square board[NUM_ROWS][NUM_COLUMNS], player currentPlayer, i
  */
 bool winCheck(player currentPlayer);
 
-void delay(int number_of_seconds);
+void delay(float number_of_seconds);
